requirements.txt
mysql-connector-python
python-tk
pyodbc